Universe 182 Crime Events Minigame Script
In this minigame, crime events will appear in random locations across the map. Players will have 5 minutes to arrive to the scene and stop the crime. If a player manages to defeat all the henchmen in time, they are awarded some gold and a mythic weapon. When the henchmen are defeated or the time runs out, another crime event will start after 5 minutes. Crime events continuously spawn throughout the lifetime of this free-for-all game mode.